refactor(util-generate-id): move into private package
