+++ shared/components/FormField/package.json
@@ -9,5 +9,8 @@
     "@tds/core-input-feedback": "^1.0.2",
     "@tds/core-paragraph": "^1.0.2",
     "@tds/core-text": "^1.0.2"
+  },
+  "devDependencies": {
+    "@tds/util-generate-id": "^1.0.0-dev"
   }
 }
