+++ shared/utils/generateId/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@tds/util-generate-id",
+  "version": "1.0.0-dev",
+  "private": true,
+  "main": "",
+  "module": ""
+}
